requirements.txt
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.30.0
Pillow>=9.0.0
tqdm>=4.64.0
ultralytics>=8.0.0        # for optional YOLOv8 usage
gradio>=3.0.0            # for demo UI 
moviepy>=1.0.0
imageio-ffmpeg>=0.4.0
scikit-learn>=1.0.0
opencv-python>=4.5.0